+++ cmd/hidden_lake/helpers/adapters/common/consumer/internal/adapted/adapted_test.go
@@ -1,4 +1,4 @@
-package common
+package adapted
 
 import "testing"