+++ cmd/hidden_lake/helpers/adapters/common/consumer/pkg/app/init_app.go
@@ -1,0 +1,23 @@
+package app
+
+import (
+	"fmt"
+	"path/filepath"
+	"strings"
+
+	"github.com/number571/go-peer/cmd/hidden_lake/helpers/adapters/common/consumer/internal/config"
+	"github.com/number571/go-peer/cmd/hidden_lake/helpers/adapters/common/consumer/pkg/settings"
+	"github.com/number571/go-peer/internal/flag"
+	"github.com/number571/go-peer/pkg/types"
+)
+
+func InitApp(pArgs []string, pDefaultPath string) (types.IRunner, error) {
+	inputPath := strings.TrimSuffix(flag.GetFlagValue(pArgs, "path", pDefaultPath), "/")
+
+	cfg, err := config.InitConfig(filepath.Join(inputPath, settings.CPathYML), nil)
+	if err != nil {
+		return nil, fmt.Errorf("init config: %w", err)
+	}
+
+	return NewApp(cfg), nil
+}