+++ cmd/hidden_lake/helpers/adapters/common/producer/internal/adapted/adapted_test.go
@@ -1,0 +1,5 @@
+package adapted
+
+import "testing"
+
+func TestNothing(_ *testing.T) {}