+++ cmd/hidden_lake/helpers/adapters/common/producer/internal/config/types.go
@@ -1,0 +1,22 @@
+package config
+
+import (
+	logger "github.com/number571/go-peer/internal/logger/std"
+	net_message "github.com/number571/go-peer/pkg/network/message"
+)
+
+type IConfig interface {
+	GetLogging() logger.ILogging
+	GetSettings() IConfigSettings
+
+	GetAddress() string
+	GetConnection() IConnection
+}
+
+type IConfigSettings interface {
+	net_message.ISettings
+}
+
+type IConnection interface {
+	GetSrvHost() string
+}